+++ x-pack/platform/plugins/shared/streams_app/public/hooks/use_stream_detail.tsx
@@ -10,7 +10,10 @@ import type { StreamsRepositoryClient } from '@kbn/streams-plugin/public/api';
 import { EuiFlexGroup, EuiLoadingSpinner } from '@elastic/eui';
 import { Streams } from '@kbn/streams-schema';
 import { STREAMS_UI_PRIVILEGES } from '@kbn/streams-plugin/public';
+import { getAncestorsAndSelf, getSegments } from '@kbn/streams-schema';
 import { useStreamsAppFetch } from './use_streams_app_fetch';
+import { useStreamsAppBreadcrumbs } from './use_streams_app_breadcrumbs';
+import { useStreamsAppParams } from './use_streams_app_params';
 import { useKibana } from './use_kibana';
 
 export interface StreamDetailContextProviderProps {
@@ -77,6 +80,24 @@ export function StreamDetailContextProvider({
     },
     [streamsRepositoryClient, name, canManage]
   );
+
+  const {
+    path: { key },
+  } = useStreamsAppParams('/{key}', true);
+
+  useStreamsAppBreadcrumbs(() => {
+    if (!definition || !Streams.WiredStream.Definition.is(definition.stream)) {
+      return [{ title: key, path: `/{key}`, params: { path: { key } } }];
+    }
+    // Build breadcrumbs for each segment in the hierarchy for wired streams
+    const ids = getAncestorsAndSelf(key);
+    const segments = getSegments(key);
+    return ids.map((id, idx) => ({
+      title: segments[idx],
+      path: `/{key}`,
+      params: { path: { key: id } },
+    }));
+  }, [key, definition]);
 
   const context = React.useMemo(
     // useMemo cannot be used conditionally after the definition narrowing, the assertion is to narrow correctly the context value
